backend/pyproject.toml
[project]
name = "nexusbase"
version = "0.1.0"
description = "NexusBase — Enterprise-grade RAG system with self-corrective retrieval"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "pymupdf>=1.24.0",
    "pillow>=10.0.0",
    "tavily-python>=0.5.0",
    "httpx>=0.27.0",
    "openai>=1.40.0",
    "google-generativeai>=0.8.0",
    "neo4j>=5.20.0",
    "redis>=5.0.0",
    "redis-om>=0.3.0",
    "arq>=0.26.0",
    "ragas>=0.2.0",
    "datasets>=3.0.0",
    "pandas>=2.0.0",
    "tabulate>=0.9.0",
    "anthropic>=0.30.0",
    "cryptography>=42.0.0",
    "dspy-ai>=2.4.0",
    "PyJWT>=2.8.0",
    "numpy>=1.26.0",
    "tiktoken>=0.7.0",
    "langchain-text-splitters>=0.3.0",
    "tenacity>=8.2.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
